backend/requirements.txt
# Production requirements - optimized for inference only
# Core FastAPI stack
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==4.6.0
aioredis==2.0.1
fastapi-limiter==0.1.5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# Payments
stripe==7.8.0

# Core utilities
python-multipart==0.0.6
httpx==0.25.2
celery==5.3.4
python-dotenv==1.0.0
aiohttp==3.9.1
ratelimit==2.2.1
nfl_data_py>=0.3.0
beautifulsoup4>=4.12.0
httpx>=0.24.0
xgboost>=2.0.0
backoff==2.2.1

# ML inference only - using CPU versions
tensorflow-cpu==2.16.1  # Much smaller than full tensorflow
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
joblib==1.3.2
scipy==1.11.4

# LLM and AI capabilities
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-community==0.0.10
openai>=1.10.0,<2.0.0
anthropic==0.8.1

# Vector database and embeddings
weaviate-client==3.25.3
sentence-transformers==2.2.2
chromadb==0.4.22

# WebSocket support for real-time features
websockets==12.0

# Enhanced caching and streaming
redis-py-cluster==2.1.3
sse-starlette==1.8.2

# Additional utilities for LLM features  
tiktoken==0.5.2
tenacity==8.2.3
asyncio-throttle==1.0.2

# Note: xgboost and lightgbm removed - models are pre-trained
# Note: visualization libraries removed - not needed for API